Move `Daemon.containerCopy` to daemon/archive.go
It's the only place where it's used.
